Update readme and changelog
I also bumped the version because you didn't already.
I use FMTK's publish feature so my mods get a version bump as soon as I push a new release.

You *might* want to bump the middle number, but that's an entirely up to feeling type thing so I'll leave it in your hands to decide.